Fix W_EXITCODE/W_STOPCODE compilation errors on MinGW and correct Cygwin platform detection (#268)
* Initial plan

* Fix W_EXITCODE/W_STOPCODE compilation and Cygwin detection

- Add fallback definitions for W_EXITCODE and W_STOPCODE macros when not available
- Keep Cygwin as Windows ABI (for calling convention) but don't define MSWINDOWS (to allow POSIX functions)
- Fixes compilation errors on MinGW/MSYS2 environments


* Skip fork-dependent tests on MinGW targets

- Add logic to skip unix_wait_helpers_demo test on MinGW (which lacks fork)
- Test still runs on Cygwin/MSYS/Unix which have POSIX fork support
- Other Unix tests (gethostname, getdomainname) continue to run on all platforms


* be better

* better

* we will never test this reliably

---------
